requirements.txt
Django>=4.1.5,<5.0.0
django-cleanup>=6.0.0,<7.0.0
django-cors-headers>=3.13.0,<4.0.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.2.2,<6.0.0
psycopg2>=2.9.0,<2.9.5
uWSGI>=2.0.19.1,<2.1
drf-spectacular>=0.26.0,<1.0.0